requirements-dev.txt
# 生产依赖
-r requirements.txt

# 开发工具
black==23.12.1
flake8==6.1.0
isort>=5.12.0,<6.0.0
mypy>=1.7.0,<2.0.0
pre-commit==3.5.0
pylint==3.0.2

# 测试框架
pytest>=7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==20.1.0
responses==0.24.0

# 文档生成
mkdocs>=1.5.3
mkdocs-material==9.5.3
mkdocstrings[python]==0.23.0

# 数据库工具
alembic>=1.12.1,<2.0.0
sqlalchemy2-stubs==0.0.2a35

# Jupyter
jupyter==1.0.0
ipython==8.17.2

# 调试工具
debugpy==1.8.0
ipdb==0.13.13

# API测试
httpie==3.2.2
locust==2.20.1

# 安全扫描
bandit==1.7.7
safety>=2.3.5

# 代码格式化
ruff>=0.1.6

# 类型检查
types-redis==4.6.0.10
types-requests==2.31.0.10
types-python-dateutil==2.8.19.14

# 数据库浏览器
pgcli==4.0.1

# 其他开发工具
watchdog==3.0.0
python-dotenv==1.0.0
pyyaml>=6.0.0
pydantic>=2.12.0
fastapi>=0.105.0